cookbook/ai-github-search/main.py
import os
import sys
from typing import Dict, Any
from fastapi import FastAPI, Query
from fastapi.responses import HTMLResponse, JSONResponse
from fastapi.staticfiles import StaticFiles
from pocketflow import Flow

# Add the current directory to Python path for imports
current_dir = os.path.dirname(os.path.abspath(__file__))
if current_dir not in sys.path:
    sys.path.insert(0, current_dir)

from flow import create_search_flow

# Try to load .env if python-dotenv is available
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

app = FastAPI(title="AI GitHub Search")
static_dir = os.path.join(os.path.dirname(__file__), "static")
app.mount("/static", StaticFiles(directory=static_dir), name="static")


@app.get("/", response_class=HTMLResponse)
def index():
    with open(os.path.join(static_dir, "index.html"), "r", encoding="utf-8") as f:
        return HTMLResponse(f.read())


@app.get("/api/search")
def api_search(q: str = Query("", min_length=1)):
    shared: Dict[str, Any] = {"nl_query": q}
    flow: Flow = create_search_flow()
    flow.run(shared)
    return JSONResponse({
        "query": q,
        "results": shared.get("results", []),
        "gh_queries": shared.get("gh_queries", []),
    })


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(app, host="127.0.0.1", port=int(os.getenv("PORT", 8000)))


