cookbook/ai-github-search/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pocketflow
PyYAML==6.0.1
jinja2==3.1.4
aiofiles==23.2.1
openai>=1.35.0 ; python_version>='3.9'
google-genai>=0.3.0 ; python_version>='3.9'

